feature: Move Payment API methods to InteractsWithPaymentApi trait
